#4 Remove debug logging
The logging is almost useless for users but a developer can debug in IDE.
The raw keys bytes are not saying anything.
We have errors to return a message to a caller.
In the same time the code needs to be compiled and it makes the program slower because it resides in CPU cache.